backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
pydantic
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-multipart
jose
requests
